# Function multiversioning for a performance boost