@Shnatsel 2024-09-29
Adds some transitive `unsafe` but I'm not worried about it because none of it is reachable with attacker-controlled data.