@fintelia 2024-09-29
I'm not seeing end-to-end decoding speed improvements. I think the cost of running runtime feature detection for every row may outweigh the time savings from faster filtering:

decode reprocessed qoi-bench images...
```
main:                  197.0 MP/s (average)  176.0 MP/s (geomean)
multiversion:          196.0 MP/s (average)  175.0 MP/s (geomean)
main-unstable:         202.7 MP/s (average)  182.9 MP/s (geomean)
multiversion-unstable: 202.5 MP/s (average)  182.7 MP/s (geomean)
```

Encoding shows only a very slight speedup...
```
main:           235.8 MP/s  28.33%
multiversion:   237.4 MP/s  28.33%
```