@Shnatsel 2024-09-29
wrt decoding: on my profile a large RGB or RGBA image with mostly Paeth filtering spends 30% of the decoding time in Paeth unfiltering, so if multiversioning introduces no overhead then the end-to-end speedup would be 4%. I am seeing exactly that speedup on a 4K RGBA image. I'm also seeing a 3% improvement on a 1024x1024 RGBA image, so the multiversioning overhead is not dramatic.

Those x86-64-v3 gains are very tantalizing! I wonder what gets vectorized and where to get us these results. How are you measuring these? I'd be very interested in reproducing these.