@Shnatsel 2024-09-29
I cannot measure any difference in end-to-end in decoding between SSE 4.1 multiversioning and `RUSTFLAGS='-C target-cpu=x86-64-v3'`. I guess it just doesn't benefit my machine for whatever reason.

On your machine are you seeing any improvement from replacing `"x86_64+sse+sse2+sse3+sse4.1+ssse3"` strings in the code with `"x86_64+sse+sse2+sse3+sse4.1+sse4.2+ssse3+avx+avx2+fma"` ? That is more or less v3.