@fintelia 2024-09-30
I've been using [my own benchmark harness](https://github.com/fintelia/corpus-bench) for measurements. So far it hasn't been intended for public consumption, so I probably need to polish the interface and expand the readme before it makes sense to anyone other than me :smile:  

 The test data was re-encoded versions of the qoi benchmark suite images, so many/most were far smaller than 1024x1024. Haven't looked at the full distribution, but at first glance I see a bunch of 64x64 icons included.

> On your machine are you seeing any improvement from replacing "x86_64+sse+sse2+sse3+sse4.1+ssse3" strings in the code with "x86_64+sse+sse2+sse3+sse4.1+sse4.2+ssse3+avx+avx2+fma" ? That is more or less v3.

Nope, I see the same performance for both.