@Shnatsel 2024-10-05
Interesting! It seems that the performance boost on your machine comes from somewhere other than filtering.

It would be interesting to profile the process with [samply](https://crates.io/crates/samply) at a high sampling rate, say `samply record -r 5000`, and compare the profiles between binaries compiled with baseline and v3 profiles. That should point us to the part that actually gets the speed boost.

As for this PR, in its current form it's probably not worth it, and without access to a machine that benefits from v3 optimizations I will not be able to work on this any further. Since there's nothing further I can do, I'm going to go ahead and close this PR.